fixed a bug where rounding would not function as expected

+++ concat_wav_files.py
@@ -10,7 +10,7 @@ files = [path + x for x in os.listdir(path) if x is not target]
 
 frames_per_second = 16000
 timeout_in_seconds = 2. * 400./1000.
-five_twelve_blocks = numpy.ceil(frames_per_second * timeout_in_seconds/512 )
+five_twelve_blocks = int(numpy.ceil(frames_per_second * timeout_in_seconds/512))
 filler_array = numpy.zeros((five_twelve_blocks*512,), numpy.int16)
 
 with sf.SoundFile('/home/rfeldhans/Music/' + target, 'w', samplerate=16000, channels=1) as target_file:
